backend/requirements.txt
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djoser==2.2.0
drf-extra-fields==3.6.1
Pillow==10.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.26.0
django-filter==23.2
python-decouple==3.5
webcolors==1.13
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
